pyproject.toml
[project]
name = "01-simple-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "bleach>=6.3.0",
    "cachetools>=6.2.2",
    "colorama>=0.4.6",
    "django>=6.0",
    "django-taggit>=6.1.0",
    "dnspython>=2.8.0",
    "markdown>=3.10",
    "pillow>=12.0.0",
    "shell>=1.0.1",
]
